Add dmux
It's a DBus signal multiplexer.   AwesomeWM, as of version 3.5.9,
doesn't allow multiple subscribers to a single DBus signal.  This
is a problem for me, because I want my widget toolkit to track
changes to my media players' statuses, but I need my keybindings to
track the same information.  dmux overrides the Awesome DBus API
so that multiple subscribers can monitor the same signal