Fix tag rules
The EWMH rules for Awesome assigns a default tag to each new client
before the manage signal is fired, and due to default window geometries
(my guess would be x=0, y=0), that default tag is always the currently
selected tag on the left-most screen.  I prefer my new windows to be
assigned to the window my mouse is currently in, so back in e7962a3
I used c:move_to_tag(mouse.screen.selected_tag) to move new windows
to the tag under the mouse.  Problem solved, right?  Nope!

The problem with this is that the manage signal handler I added this
to runs *after* rules.rules, so any tags assigned via rules are
overwritten.  The workaround is to check to see if the currently
assigned tag is the default one - if so, move it to the proper screen.
I thought about this, and I *think* this should work ok on both single
and multi monitor setups.