Remove client rule for Firefox
I don't really like how when I open new windows, they are automatically
routed to the first tag.  Rather than try to figure out a clever
strategy for when this _should_ happen, just remove the rule.  I really
only care about this for the first Firefox window anyway, and I rarely
restart Firefox as it is